backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
groq>=0.4.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
python-jobspy>=1.1.0
lxml>=4.9.0
supabase>=2.0.0
googlesearch-python>=1.2.0
